+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * enhancements
   * [List] Add support to keyreplace and keystore
   * [Kernel] Better error messages when invalid options are given to `import`, `alias` or `require`
+  * [Binary] Add support to `part` and `split` functions
 
 # v0.6.0 (2012-08-01)